Refactor Django app structure
- Add verbose names for all apps
- Move general stuff into core app
