+++ integreat_cms/api/apps.py
@@ -2,6 +2,7 @@
 Meta information about the Integreat CMS API app
 """
 from django.apps import AppConfig
+from django.utils.translation import gettext_lazy as _
 
 
 class ApiConfig(AppConfig):
@@ -9,4 +10,7 @@ class ApiConfig(AppConfig):
     Configuration parameters for the API app
     """
 
+    #: Full Python path to the application
     name = "integreat_cms.api"
+    #: Human-readable name for the application
+    verbose_name = _("API")
