+++ integreat_cms/core/signals/__init__.py
@@ -1,0 +1,5 @@
+"""
+This package contains signal handlers (see :doc:`django:topics/signals`).
+"""
+from . import auth_signals
+from . import feedback_signals
