+++ integreat_cms/core/signals/feedback_signals.py
@@ -1,9 +1,13 @@
+"""
+This module contains signal handlers related to feedback objects.
+"""
+
 from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
 
 from cacheops import invalidate_obj
 
-from ..models import (
+from ...cms.models import (
     Feedback,
     EventFeedback,
     EventListFeedback,
