+++ integreat_cms/locale/de/LC_MESSAGES/django.po
@@ -18,6 +18,14 @@ msgstr ""
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1);\n"
 
+#: api/apps.py
+msgid "API"
+msgstr "API"
+
+#: cms/apps.py
+msgid "CMS"
+msgstr "CMS"
+
 #: cms/constants/administrative_division.py cms/templates/pois/poi_list.html
 #: cms/templates/pois/poi_list_archived.html
 msgid "City"
@@ -8180,6 +8188,10 @@ msgid "Only superusers can enable or disable passwordless authentication."
 msgstr ""
 "Nur Administrator:innen können die Passwortlose Authentifizierung aktivieren"
 
+#: core/apps.py
+msgid "Core"
+msgstr "Core"
+
 #: core/settings.py
 msgid "German"
 msgstr "Deutsch"
@@ -8191,6 +8203,10 @@ msgstr "Englisch"
 #: core/settings.py
 msgid "Dutch"
 msgstr "Niederländisch"
+
+#: deepl_api/apps.py
+msgid "DeepL API"
+msgstr "DeepL API"
 
 #: deepl_api/utils.py
 msgid "No source translation could be found for {} \"{}\"."
@@ -8212,6 +8228,18 @@ msgstr "{} \"{}\" wurde erfolgreich übersetzt"
 msgid "{} \"{}\" could not be translated automatically."
 msgstr "{} \"{}\" konnte nicht automatisch übersetzt werden."
 
+#: firebase_api/apps.py
+msgid "Firebase API"
+msgstr "Firebase API"
+
+#: gvz_api/apps.py
+msgid "GVZ API"
+msgstr "GVZ (Gemeindeverzeichnis) API"
+
+#: matomo_api/apps.py
+msgid "Matomo API"
+msgstr "Matomo API"
+
 #: matomo_api/matomo_api_client.py
 msgid "Total Accesses"
 msgstr "Alle Zugriffe"
@@ -8227,6 +8255,18 @@ msgstr "WebApp Zugriffe"
 #: matomo_api/matomo_api_client.py
 msgid "CW"
 msgstr "KW"
+
+#: nominatim_api/apps.py
+msgid "Nominatim API"
+msgstr "Nominatim API"
+
+#: sitemap/apps.py
+msgid "Sitemap"
+msgstr "Sitemap"
+
+#: summ_ai_api/apps.py
+msgid "SUMM.AI API"
+msgstr "SUMM.AI API"
 
 #: summ_ai_api/utils.py
 msgid "No German translation could be found for {} \"{}\"."
@@ -8245,6 +8285,14 @@ msgstr ""
 #: summ_ai_api/utils.py
 msgid "{} \"{}\" has been successfully translated into Easy German."
 msgstr "{} \"{}\" wurde erfolgreich in Leichte Sprache übersetzt."
+
+#: textlab_api/apps.py
+msgid "TextLab API"
+msgstr "TextLab API"
+
+#: xliff/apps.py
+msgid "XLIFF"
+msgstr "XLIFF"
 
 #: xliff/utils.py
 msgid ""
@@ -8347,6 +8395,9 @@ msgstr ""
 
 #~ msgid "Currently not visible in apps"
 #~ msgstr "Wird noch nicht in den Apps angezeigt"
+
+#~ msgid "Integreat CMS core"
+#~ msgstr "Integreat CMS core"
 
 #~ msgid "Edit language tree node"
 #~ msgstr "Sprach-Knoten bearbeiten"
