+++ integreat_cms/matomo_api/apps.py
@@ -1,6 +1,7 @@
 import logging
 
 from django.apps import AppConfig
+from django.utils.translation import gettext_lazy as _
 
 logger = logging.getLogger(__name__)
 
@@ -10,4 +11,7 @@ class MatomoApiConfig(AppConfig):
     MatomoApiClient config inheriting the django AppConfig
     """
 
+    #: Full Python path to the application
     name = "integreat_cms.matomo_api"
+    #: Human-readable name for the application
+    verbose_name = _("Matomo API")
