+++ integreat_cms/sitemap/apps.py
@@ -1,4 +1,5 @@
 from django.apps import AppConfig
+from django.utils.translation import gettext_lazy as _
 
 
 class SitemapConfig(AppConfig):
@@ -6,4 +7,7 @@ class SitemapConfig(AppConfig):
     This is the basic configuration for the sitemap app
     """
 
+    #: Full Python path to the application
     name = "integreat_cms.sitemap"
+    #: Human-readable name for the application
+    verbose_name = _("Sitemap")
